+++ classes/Collabot.class.js
@@ -6,6 +6,7 @@ function Collabot(config){
 	this.guid = null;
 }
 
+//TODO: This should be moved out of global scope
 function guid() {
   function s4() {
     return Math.floor((1 + Math.random()) * 0x10000).toString(16).substring(1);
@@ -15,14 +16,22 @@ function guid() {
 
 module.exports = Collabot;
 
+// Managers 
 var CollaborationManager = require('./CollaborationManager.class');
 var ConversationManager = require('./managers/ConversationManager.class');
-var DefaultConversationHandler = require('./conversationHandlers/DefaultConversationHandler.class');
-var CommandConversationHandler = require('./conversationHandlers/CommandConversationHandler.class');
-var GreetingConversationHandler = require('./conversationHandlers/GreetingConversationHandler.class');
+var TrainingSessionManager = require('./managers/TrainingSessionManager.class');
+
+// Conversation Handlers
+var DefaultConversationHandler = require('./conversationHandlers/DefaultConversationHandler.class')
+var CommandConversationHandler = require('./conversationHandlers/CommandConversationHandler.class')
+var GreetingConversationHandler = require('./conversationHandlers/GreetingConversationHandler.class')
 var CreateTrainingSessionConversationHandler = require('./conversationHandlers/CreateTrainingSessionConversationHandler.class');
-var async = require('async');
+
+// Utility Classes
 var mentionCheck = require('./util/ChatUtils.class');
+
+// External Libs
+var async = require('async');
 
 Collabot.prototype = {
 	start: function(callback){
@@ -33,6 +42,7 @@ Collabot.prototype = {
 			this.connector.init(this);
 			this.collaborationManager = new CollaborationManager();
 			this.conversationManager = new ConversationManager(this);
+			this.trainingSessionManager = new TrainingSessionManager(this);
 			this.commandConversationHandler = new CommandConversationHandler(this);
 			this.defaultConversationHandler = new DefaultConversationHandler(this);
 			this.handlers = {
@@ -97,7 +107,7 @@ Collabot.prototype = {
 		this.connector.share(text);
 	},
 	registerPlayers: function(players){
-		console.log("registering players...");
+		console.log("Registering players...");
 		console.log(players);
 		var that = this;	
 		async.each(players, 
@@ -107,24 +117,24 @@ Collabot.prototype = {
 			});		
 	},
 	registerPlayer: function(player){
-		console.log("attempting to register player: " + player);		
+		console.log("Attempting to register player: " + player);		
 		var that = this;
 		that.persistence.getPlayerByName(player, function(user, err){
 			if(err){
-				console.log("error getting player");
+				console.log("Error getting player");
 				console.error(err);
 			} else if(!user){
-				console.log("new Player: " + player);				
+				console.log("New Player: " + player);				
 				that.persistence.insertNewPlayer(player, function(inserted, err){
 					if(err){
 						console.log(err.stack);
 					} else {
-						console.log("player inserted: ");
+						console.log("Player inserted: ");
 						console.log(inserted);
 					}
 				});
 			} else {
-				console.log(user + " already exists.");
+				console.log(user.name + " already exists!");
 			}
 		});		
 	}