+++ classes/conversationHandlers/CommandConversationHandler.class.js
@@ -15,12 +15,14 @@ CommandConversationHandler.prototype = {
 		} else	if (text.indexOf("creator") > -1){
 			this._creator();
 		} else	if (text.indexOf("top") > -1){
-			this._top();
+			this._top(text);
 		} else if (text.toLowerCase().indexOf("how am i") > -1){
 			this._howAmIDoing(from);
+		} else if (/authorize (.+) as presenter$/.exec(text)){
+			this._autorizeAsPresenter(from, text);
 		} else if(text.indexOf("create training") > -1){
 			this._createTraining(from);
-		}else {
+		} else {
 			this._wtf(from);
 		}
 	},
@@ -47,8 +49,29 @@ CommandConversationHandler.prototype = {
 		console.log(updateScoreRequest)
 		this.bot.collaborationManager.givePoints(updateScoreRequest, this.bot);
 	},
-	_top: function(){
-		this.bot.collaborationManager.topTen(this.bot);
+	_top: function(text){
+		var filters = /top( day| week| month| year)*(.+)*$/.exec(text);
+		var period = null;
+		var channel = null;
+		if(filters){
+			var filter1 = filters[1]? filters[1].trim() : null;
+			var filter2 = filters[2]? filters[2].trim() : null;
+			var period = filter1;
+			if(!filter1 && filter2){
+				if(filter2 == "day" || filter2 == "week" || filter2 == "month" || filter2 == "year" ){
+					period = filter2;
+				}
+				else{
+					channel = filter2;
+				}
+			}
+			else{
+				channel = filter2;
+			}
+		}
+		
+			
+		this.bot.collaborationManager.topTen(period, channel, this.bot);
 	},
 	_howAmIDoing: function (from){
 		this.bot.collaborationManager.tellStatusTo(from, this.bot);
@@ -65,10 +88,24 @@ CommandConversationHandler.prototype = {
 	_wtf: function(who){
 		this.bot.share("Perhaps you need to rephrase... ");
 	},
+	_autorizeAsPresenter: function (from, text){
+		var command = /authorize (.+) as presenter$/.exec(text);
+		if (!command || !command.length || !command.length == 2){
+			this.bot.share("Sorry, I didn't understand that..");
+			return;
+		}
+		var presenter = command[1];
+		if (!presenter){
+			this.bot.share("Sorry, I didn't understand that..");
+			return;
+		}
+		this.bot.trainingSessionManager.requestAuthorizationAsPresenter(from, presenter);
+	},
 	_help: function (who){
 		this.bot.say(who, "["+this.bot.config.botName+" give] Gives a player X points. Example: 'bot give 5 points to slash'.");
 		this.bot.say(who, "["+this.bot.config.botName+" about] Gets some information about the collabot.");
 		this.bot.say(who, "["+this.bot.config.botName+" how am i] Tells you your overall, daily, weekly and last week scores.");
+		this.bot.say(who, "["+this.bot.config.botName+" top [day|week|month|year] [channel_name]] Tells you the top ten collaborators by period and channel name. Period and Channel are optional.");
 	},
 	_createTraining: function(from){
 		var handler = this;