+++ config.js
@@ -5,18 +5,28 @@ var MongoConnector = require("./classes/persistence/MongoConnector.class");
 
 // Slack API config karmabot for tests
 module.exports = {
-	version: "1.1",
+	version: "1.5",
 	environment: "DEVELOPMENT",
-	channel: "karma",
+	channel: "gambit-test",
 	botName: "collabot",
 	token: 'xoxb-3558335570-MXw76pJPGbNLLyLi1KYWK8qr',
+	// token: 'xoxb-3749826050-mKrPHbMYqRDa0mb47lhf3sfr', // Rick's tal-bot 
 	autoReconnect: true,
 	autoMark: true,
 	connector: SlackConnector,
 	persistence: MongoConnector,
-	//dbURL: 'mongodb://localhost:27017/collabot',
-	dbURL: 'mongodb://'+process.env.MDBUSERNAME+':'+process.env.MDBPWD+'@'+process.env.MDBHOST+':'+process.env.MDBPORT+'/collabot',
-	maxCollabPoints : 10
+	dbURL: 'mongodb://'+process.env.MDBUSERNAME+':'+process.env.MDBPWD+'@'+process.env.MDBHOST+':'+process.env.MDBPORT+'/'+process.env.MDBDB,
+	// dbURL: 'mongodb://localhost:27017/collabot',
+	appUrls: {
+		start: process.env.APPSTART, 
+			// 'http://salty-inlet-8617.herokuapp.com/start',
+			// 'http://localhost:3000/start',
+		stop: process.env.APPSTOP
+			// 'http://salty-inlet-8617.herokuapp.com/stop'
+			// 'http://localhost:3000/stop'
+	},
+	maxCollabPoints : 10,
+	trainingSessionManager: 'slizarazo'
 };
 
 // Slack API config prod