+++ Changelog.md
@@ -1,3 +1,12 @@
+
+What's New in v. 1.5
+====================
+
+- Filters for 'top' operation are available: Bot top now takes two optional arguments: a period of time (day, week, month, year) and a channel. Those values are used to filter top scores data.
+- web tier now finds the app endpoints using config file or environment properties (appUrls)
+- minor bug fixes
+
+
 What's New in v. 1.1
 ====================